cache_directory no longer throws if the directory does not exist.
This is because the directory (and any of its subdirectories) should already be
auto-populated by the relevant functions when they need to save each file.

Also, return the _previous_ location of the cache when setting a new value,
allowing callers to easily reset it, e.g., in a finally clause.